shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared code for cat flap project"
requires-python = ">=3.13"
dependencies = [
    "pyzmq",
    "Pillow",
    "numpy",
]

[tool.setuptools]
py-modules = []
